feat: Add path alias for resources/scripts
This commit adds a new configuration to the tsconfig.json file. It includes a "paths" property that maps the alias "@/*" to the "./resources/scripts/*" directory. This allows for easier import statements in the code.

The purpose of this change is to improve the organization and readability of the code by using path aliases.